# api/openapi.yaml
openapi: 3.0.2
info:
  description: |
    PayGate is a RESTful API enabling first-party Automated Clearing House ([ACH](https://en.wikipedia.org/wiki/Automated_Clearing_House)) transfers to be created without a deep understanding of a full NACHA file specification. First-party transfers initiate at an Originating Depository Financial Institution (ODFI) and are sent off to other Financial Institutions.

    Tenants are the largest grouping in PayGate and are typically a vendor who is reselling ACH services or a company making ACH payments themselves. A legal entity is linked off a Tenant as the primary Customer used to KYC and in transfers with the Tenant itself.

    An Organization is a grouping within a Tenant which typically represents an entity making ACH transfers. These include clients of an ACH reseller or business accepting payments over ACH. A legal entity is linked off an Organization as the primary Customer used to KYC and in transfers with the Organization itself.

    ![](https://raw.githubusercontent.com/moov-io/paygate/master/docs/images/tenant-in-paygate.png)

    There are also [admin endpoints](https://moov-io.github.io/paygate/admin/) for back-office operations.
  version: v1
  title: Paygate API
  contact:
    url: https://github.com/moov-io/paygate
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8082
    description: Local development
...
paths:
...
  /micro-deposits:
    post:
      tags: [Validation]
      summary: Initiate micro-deposits
      description: Start micro-deposits for a Destination to validate.
      operationId: initiateMicroDeposits
      parameters:
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMicroDeposits'
      responses:
        '200':
          description: Initiated micro-deposits for external account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroDeposits'
        '400':
          description: Problem initiating micro-deposits, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /micro-deposits/{microDepositID}:
    get:
      tags: [Validation]
      summary: Get micro-deposit information
      description: Retrieve the micro-deposits information for a specific microDepositID
      operationId: getMicroDeposits
      parameters:
        - name: microDepositID
          in: path
          description: Identifier for micro-deposits
          required: true
          schema:
            type: string
            example: c336f57e
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      responses:
        '200':
          description: micro-deposits for external account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroDeposits'
        '400':
          description: Problem reading micro-deposits, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /accounts/{accountID}/micro-deposits:
    get:
      tags: [Validation]
      summary: Get micro-deposits for a specified accountID
      description: Retrieve the micro-deposits information for a specific accountID
      operationId: getAccountMicroDeposits
      parameters:
        - name: accountID
          in: path
          description: accountID identifier from Customers service
          required: true
          schema:
            type: string
            example: c336f57e
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      responses:
        '200':
          description: micro-deposits for external account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroDeposits'
        '400':
          description: Problem reading micro-deposits, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  # Transfers
...
components:
  schemas:
...
    CreateMicroDeposits:
      properties:
        destination:
          $ref: '#/components/schemas/Destination'
      required:
        - destination
    MicroDeposits:
      properties:
        microDepositID:
          type: string
          example: 8e8cc27b
          description: A microDepositID to identify this set of credits to an external account
        transferIDs:
          type: array
          items:
            type: string
            example: d2376d77
          description: An array of transferID values created from this micro-deposit
        destination:
          $ref: '#/components/schemas/Destination'
        amounts:
          type: array
          items:
            $ref: '#/components/schemas/Amount'
        status:
          $ref: '#/components/schemas/TransferStatus'
        processedAt:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
          nullable: true
        created:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
      required:
        - microDepositID
        - transferIDs
        - destination
        - amounts
        - status
        - created
...
    Destination:
      description: Customer that is receiving a Transfer
      properties:
        customerID:
          type: string
          example: 11ffa67d
          description: A customerID from the Customers service used as source for this Transfer
        accountID:
          type: string
          example: 68b534b7
          description: A accountID from the Customers service under the specified Customer used for this Transfer. If the Customer only has one account this value can be left empty.
      required:
        - customerID
        - accountID
    Amount:
      type: string
      example: "USD 99.99"
      description: Amount of money. USD - United States.
...
    TransferStatus:
      type: string
      description: Defines the state of the Transfer
      enum:
        - canceled
        - failed
        - reviewable
        - pending
        - processed
...